api: continued work to reduce repetition by making all processing functions reusable across anime/manga queries
