+++ site/api/anime/main.py
@@ -1,7 +1,14 @@
-import pandas as pd
 from api.anime.insights import general_insights, genre_insights
-from api.anime.processing import check_nulls, get_anime_info, get_id, get_user_data
-from api.funcs import create_genre_data, create_plot_data, create_table, round_scores
+from api.funcs import (
+    check_nulls,
+    create_genre_data,
+    create_plot_data,
+    create_table,
+    get_format_info,
+    get_id,
+    get_user_data,
+    round_scores,
+)
 from api.upload import blob_upload
 
 
@@ -9,13 +16,19 @@ def fetch_anime(username: str):
     # Local testing
     # username = "keejan"
 
+    format = "anime"
+
     # NOTE: Processing
     anilist_id = get_id(username=username)
     user_score, user_info, id_list = get_user_data(
-        username=username, anilist_id=anilist_id
+        username=username,
+        anilist_id=anilist_id,
+        format=format,
     )
-    anime_info = get_anime_info(username=username, id_list=id_list)
-    anime_info, user_score = check_nulls(anime_info=anime_info, user_score=user_score)
+    anime_info = get_format_info(username=username, id_list=id_list, format=format)
+    anime_info, user_score = check_nulls(
+        format_info=anime_info, user_score=user_score, format="anime"
+    )
 
     # NOTE: Insights
     merged_dfs = user_score.merge(anime_info, on="anime_id", how="left")
